requirements.txt
Flask==0.11.1

cf-deployment-tracker==1.0.4

cloudant==2.4.0

python-keystoneclient==1.7.1

python-swiftclient==2.6.0
python-gnupg